ix: Change submitScore parameter from Int to Long to support large scores
Android's leaderboard API supports 64-bit scores (Long), but the plugin
was using 32-bit Int which caused overflow for scores > 2.1 billion.

This change:
- Updates submitScore() to accept Long instead of Int in LeaderboardsProxy.kt
- Updates signal bridge to use Long in GodotAndroidPlugin.kt
- Removes unnecessary .toLong() conversion since parameter is already Long

+++ plugin/src/main/java/com/jacobibanez/plugin/android/godotplaygameservices/GodotAndroidPlugin.kt
@@ -213,7 +213,7 @@ class GodotAndroidPlugin(godot: Godot) : GodotPlugin(godot) {
      * @param score The raw score value. For more details, please [see this page](https://developers.google.com/games/services/common/concepts/leaderboards).
      */
     @UsedByGodot
-    fun submitScore(leaderboardId: String, score: Int) =
+    fun submitScore(leaderboardId: String, score: Long) =
         leaderboardsProxy.submitScore(leaderboardId, score)
 
     /**
